Rebirth PCM API
More like how I intended it back with the mixer patch.

NOTE: WILL NOT COMPILE!!! HERE FOR PERUSAL ONLY AT THIS TIME!!!

It's fairly close though. I just want it backed up here and add
some visibility.

* Some unifications of recording and playback functions

* Mono and stereo support for all targets

* Framework for multiple source buffer formats

* Framework for output of more that 16-bits per sample

* Initially, driver DMA depth will be statically configured

I'll possibly just merge in the sample-sizes-to-frame-counts
patch since the changes here immediately follow it and make
drastic changes to that change.

+++ firmware/export/pcm_sampr.h
@@ -373,8 +373,8 @@ extern const unsigned long rec_freq_sampr[REC_NUM_FREQ];
 #endif
 
 #ifndef PCM_SAMPR_CONFIG_ONLY
-unsigned int pcm_sampr_to_hw_sampr(unsigned int samplerate,
-                                   unsigned int type);
+unsigned long pcm_sampr_to_hw_sampr(unsigned long samplerate,
+                                    unsigned int type);
 #endif
 
 #else /* ndef CONFIG_SAMPR_TYPES */
